pkg: SetDefaults, Validate, Flavor
This commit uses the `SetDefaults` method to set default values on any
Notebook resource that will be reconciled. It also uses the `Validate`
method to ensure that notebooks to be reconciled are valid and bails
otherwise. Finally, it introduces a `Flavor` field on the `NotebookSpec`
that allows selecting the base image for the Jupyter notebook pod.